+++ src/generators/fragments.generator.ts
@@ -1,74 +1,60 @@
 import { Type, Field, OfType } from '../models';
-import { evaluateType } from '../utilities';
-import { range } from 'lodash';
+import { SchemaStore } from '@store';
 
-export const createConnectionFragment = (
-  typeName: string,
-  allTypes: Type[],
-  fragment: string
-): string | null => {
-  const foundType = allTypes.find((f) => f.name === typeName);
+/** Generate a connection fragment for fields like `pageInfo`, and a placeholder for insertable content */
+export function generateConnectionFragment(typeName: string, fragment: string): string | null {
+  const foundType = SchemaStore.findType(typeName);
   if (foundType) {
     let outputFragment = '';
 
-    const createLines = (field: Field) => {
+    const generateLines = (field: Field): void => {
       outputFragment += `${field.name} `;
-      const { typeName, isScalar } = evaluateType(field);
+      const { fieldName, isScalar } = SchemaStore.getFieldProps(field);
       if (!isScalar) {
         if (field.name === 'node') {
           outputFragment += `{${fragment}} `;
         } else {
           outputFragment += `{ `;
-          const type = allTypes.find((f) => f.name === typeName);
-          type?.fields?.forEach(createLines);
+          const type = SchemaStore.findType(fieldName);
+          type?.fields?.forEach(generateLines);
           outputFragment += `} `;
         }
       }
     };
-    foundType.fields.map(createLines);
+    foundType.fields.map(generateLines);
     return outputFragment;
   }
   return null;
-};
+}
 
-export const createNormalFragment = (
-  type: Type,
-  allTypes: Type[],
-  connection?: boolean
-): { fragment: string; foundName: string } | null => {
-  let foundType = allTypes.find((f) => f.name === type.name);
-  if (foundType) {
-    let outputFragment = '';
-    const THRESHOLD = connection ? 4 : 2;
+/** Generate a normal Type fragment usable everywhere */
+export function generateNormalFragment(type: Type, connection?: boolean): string | null {
+  let outputFragment = '';
+  const THRESHOLD = connection ? 4 : 2;
 
-    const createFragmentWithDeps = (field: Field, level: number) => {
-      let { typeName, isScalar, isEnum, isEdge } = evaluateType(field);
-      if (!isScalar && !isEnum) {
-        const type = allTypes.find((f) => f.name === typeName);
-        if (level < THRESHOLD) {
-          outputFragment += `${field.name} {`;
-          type?.fields?.forEach((m) => createFragmentWithDeps(m, level + 1));
-          outputFragment += `} `;
-        } else if (type?.fields && type.fields.length < 7) {
-          outputFragment += `${field.name} {`;
-          type?.fields?.forEach((m) => createFragmentNoDeps(m, level + 1));
-          outputFragment += `} `;
-        }
-      } else {
-        outputFragment += `${field.name} `;
+  const createFragmentWithDeps = (field: Field, level: number): void => {
+    let { fieldName, isScalar, isEnum } = SchemaStore.getFieldProps(field);
+    if (!isScalar && !isEnum) {
+      const type = SchemaStore.findType(fieldName);
+      if (level < THRESHOLD) {
+        outputFragment += `${field.name} {`;
+        type?.fields?.forEach((field) => createFragmentWithDeps(field, level + 1));
+        outputFragment += `} `;
+      } else if (type?.fields && type.fields.length < 7) {
+        outputFragment += `${field.name} {`;
+        type?.fields?.forEach((field) => createFragmentNoDeps(field));
+        outputFragment += `} `;
       }
-    };
-    const createFragmentNoDeps = (field: Field, level: number) => {
-      let { isScalar, isEnum } = evaluateType(field);
-      if (isScalar || isEnum) {
-        outputFragment += `${field.name} `;
-      }
-    };
-    foundType.fields.map((m) => createFragmentWithDeps(m, 0));
-    return {
-      fragment: outputFragment,
-      foundName: foundType.name,
-    };
-  }
-  return null;
-};
+    } else {
+      outputFragment += `${field.name} `;
+    }
+  };
+  const createFragmentNoDeps = (field: Field) => {
+    let { isScalar, isEnum } = SchemaStore.getFieldProps(field);
+    if (isScalar || isEnum) {
+      outputFragment += `${field.name} `;
+    }
+  };
+  type.fields.map((m) => createFragmentWithDeps(m, 0));
+  return outputFragment;
+}