+++ src/generators/index.ts
@@ -1,6 +1,5 @@
 export * from './fragments.generator';
-export * from './methods.generator';
+export * from './functions.generator';
 export * from './templates.generator';
 export * from './types.generator';
 export * from './hooks.generator';
-export * from './models.generator';