+++ src/index.ts
@@ -1,3 +1,3 @@
 export { sgtsGenerate } from './runtime';
-export { fetchSchemas } from './utilities';
+export { retrieveIntrospectionSchema } from './utilities';
 export { SgtsConfig } from './models';