+++ src/templates/imports.template.ts
@@ -1,20 +1,10 @@
 import { guessFragmentTypeTemplate } from './fragmentType.template';
+import { ParametersStore } from '@store';
 
-interface DefineImportsArgs {
-  codegenMethods: boolean;
-  codegenReactHooks: boolean;
-  codegenTemplates: boolean;
-  codegenVueHooks: boolean;
-}
-
-export const defineImports = ({
-  codegenReactHooks,
-  codegenVueHooks,
-  codegenMethods,
-  codegenTemplates,
-}: DefineImportsArgs): string => {
+export const defineImports = (): string => {
   let template = '';
-  if (codegenReactHooks || codegenMethods || codegenTemplates || codegenVueHooks) {
+  const { isCodeGen, codegenFunctions, codegenVueHooks, codegenReactHooks } = ParametersStore;
+  if (isCodeGen) {
     template = `
     import { OperationDefinitionNode } from 'graphql';
     ${guessFragmentTypeTemplate}
@@ -31,7 +21,7 @@ export const defineImports = ({
       export type ReactiveFunction<TParam> = () => TParam;
       `;
     }
-    if (codegenMethods) {
+    if (codegenFunctions) {
       template += `
       import ApolloClient, { QueryOptions, OperationVariables, MutationOptions, ObservableQuery } from 'apollo-client';
       import { execute } from 'apollo-link';