+++ src/templates/index.ts
@@ -4,3 +4,4 @@ export * from './withReactHooksTemplate';
 export * from './withVueHooks.template';
 export * from './imports.template';
 export * from './config.template';
+export * from './shared.template';