+++ src/utilities/provider.service.ts
@@ -4,11 +4,13 @@ import ora from 'ora';
 import * as query from 'querystringify';
 import { getIntrospectionQuery, printSchema } from 'graphql';
 import { buildClientSchema } from 'graphql/utilities/buildClientSchema';
-import { GraphQLJSONSchema } from '../models';
+import { GraphQLJSONSchema, ObjectLiteral } from '../models';
 import path from 'path';
 import fs from 'fs';
 
-export const downloadSchema = async (endpoint: string, header?: string): Promise<string> => {
+const possibleGraphQLSuffix = ['/graphql', '/api', '/__graphql', '/__api'];
+
+export async function downloadSchema(endpoint: string, header?: string): Promise<string> {
   const download = ora(`⬇️  Downloading schemas from ${chalk.blue(endpoint)}`).start();
   try {
     let formatedHeaders = getHeadersFromInput(header);
@@ -39,21 +41,17 @@ export const downloadSchema = async (endpoint: string, header?: string): Promise
       `
     );
   }
-};
+}
 
-function getHeadersFromInput(header: any): { [key: string]: string } {
+function getHeadersFromInput(header?: string | ObjectLiteral): { [key: string]: string } {
   switch (typeof header) {
     case 'string': {
-      const keys = query.parse(header);
+      const keys = query.parse(header) as ObjectLiteral;
       const key = Object.keys(keys)[0];
       return { [key]: keys[key] };
     }
     case 'object': {
-      return header.map((header) => {
-        const keys = query.parse(header);
-        const key = Object.keys(keys)[0];
-        return { [key]: keys[key] };
-      });
+      return header;
     }
     default: {
       return {};
@@ -99,9 +97,7 @@ async function getRemoteSchema(
   }
 }
 
-const possibleGraphQLSuffix = ['/graphql', '/api', '/__graphql', '/__api'];
-
-export async function fetchSchemas({
+export async function retrieveIntrospectionSchema({
   endpoint,
   header,
   json,
@@ -114,8 +110,8 @@ export async function fetchSchemas({
 }): Promise<GraphQLJSONSchema | null> {
   try {
     if (endpoint) {
-      const graphqlRegxp = /[^/]+(?=\/$|$)/;
-      const [result] = graphqlRegxp.exec(endpoint);
+      // const graphqlRegxp = /[^/]+(?=\/$|$)/;
+      // const [result] = graphqlRegxp.exec(endpoint);
       const JSONschema = await downloadSchema(endpoint, header);
       if (download) {
         const outputfile = path.resolve(process.cwd(), download);