+++ types/builders/methods.builder.d.ts
@@ -1,8 +1,2 @@
-import { Field, GraphQLJSONSchema, CodeGenType } from '../models';
-interface CreateMethodsArgs {
-    schema: GraphQLJSONSchema;
-    mode?: CodeGenType;
-}
-export declare const createMethods: ({ schema, mode }: CreateMethodsArgs) => Promise<string>;
-export declare function retrieveQueriesList({ schema }: CreateMethodsArgs): [Field[], Field[], Field[]];
-export {};
+import { CodeGenType } from '@models';
+export declare function createMethods(mode: CodeGenType): string;