+++ types/generators/hooks.generator.d.ts
@@ -1,9 +1,9 @@
-import { Field, MethodType } from '../models';
+import { Field, MethodType } from '@models';
 interface GraphQLFunctionArgs {
     field: Field;
-    type: MethodType;
-    renderedFragmentInner: string;
+    functionType: MethodType;
+    innerFragment: string;
 }
-export declare const createReactApolloHook: ({ field, type, renderedFragmentInner, }: GraphQLFunctionArgs) => string;
-export declare const createVueApolloHook: ({ field, type, renderedFragmentInner, }: GraphQLFunctionArgs) => string;
+export declare const createReactApolloHook: ({ field, functionType, innerFragment, }: GraphQLFunctionArgs) => string;
+export declare const createVueApolloHook: ({ field, functionType, innerFragment, }: GraphQLFunctionArgs) => string;
 export {};